refactor products from index into read-write methods on Experiment
